Lesstif + GTK HIDs: Ignore ISO Level3 shift modifier key
This key press is found on some international keyboards (e.g. Spanish).
If we don't ignore it, PCB complains about the key not being tied to
an action when it is pressed.

Thanks to Gabriel for reporting this and providing the patch for
Lesstif. I have made the equivalent change to GTK as well.

Reported-By: Gabriel Paubert <paubert@iram.es>
Signed-off-By: Peter Clifton <pcjc2@cam.ac.uk>