Try snapping to element marks first, as they are only a weak snap.
Because the element mark is never chosen in preference to a closer grid
point, we should try snapping to it first - rather than last.

Currently, we might have a crosshair location where (say), a pad is snapped
to in preference to the nearest grid-point, then the element mark is tested
and snapped to because it is closer than the pad. This can occur even when
there is a closer grid point than the element mark, since the grid snap
(nearest_is_grid == true) has already been discarded by the snap to the pad.