crosshair.c: Snap to points along off-grid lines when drawing tracks
This should greatly easy making tidy layouts where some lines have
(perhaps by necessity) ended up off-grid.

This patch adds code to snap onto the center of a line. It finds
the nearest grid point to the cursor, then will allow snapping at
the intersections between the line in question and the lines of an
imaginary X and + centered on the nearest grid-point to the cursor.

This allows neat drawing of horizontal, vertical and 45 degree lines
which will land correctly on the existing line.