line.c: Fix a bogus compiler warning regarding a X + n >= X being always true.
It would appear that gcc (4.6.1-5ubuntu1) is inoring the full definition
of INDEXOFCURRENT and fixating on portions of it which will indeed always
evaluate to a particular value.

Since these are conditional, e.g.: (cond ? a : b), and there appears to be
no reason for GCC to assume the controlling condition is always true, it
is "probably" safe to assume that this warning manifests due to gcc
considering each portion of the INDEXOFCURRENT definition in turn.

It appears we can avoid this particular warning by temporarily assigning
INDEXOFCURRENT into a variable before we test with it.