hid/ps: Fix arc drawing since pcb_printf patch
The last parameter to the arc macro we use for drawing arcs is unitless,
so should be a double. This was broken by:

commit fa9ae1f6b1eabaab961795ce7be53afe46eaa735

    Convert ps/eps/lpr HID's to use pcb-printf

I've reverted the change, and added an extra (double) cast compared
to what was there before, I'm not sure we wouldn't get an integer
division without it. One day I should learn these type promotion /
casting rules properly! ;)