DrawLayer(): Pass clear_pins argument as true when calling DrawLayerCommon
This should not affect any rendering. If the check planes flag is set, we
should only draw polygons, nothing more. Only a handful of cases draw layers
via this code-path, and in all but the exporter cases, they are not called
if CHECKPLANES is set.

The exporters which call DrawLayer explicitly clear the CHECKPLANES flag
(since commit c00ee74844cad14ecd55bae854a82ff6412633e6), so nothing should
be affected by this argument change.