hid/gtk: Attempt to fix logic to flip component / solder group visibility on flip
Should get back to better behaviour. I probably broke this somewhat with
commit f903b4be6b85efc110852f7be40edf8245f0a513, which attempted to
re-state the previous logic in a clearer fashon.

The logic now should:

If flipping sides, and only ONE of the solder / component layers (groups)
is visible, and that layer (group) is _active_, then swap the visibilities
of the component / solder layers (groups), and make the newly visible
layer (group) active.

There are still bugs in this code relating to the assumption that the
first entry in the layer group is the one which is being toggled. This
breaks if the first entry in the group is that corresponding to the silk
for that side of the board.