hid/gtk: Don't use a GtkEventBox as a container widget
We weren't making use of the events from the event box, and appeared
just to be using it as a convenient container to allow setting the
sensitivity on our entire left hand toolbar at once.

Scrap the event box widget and just store a pointer to the vbox widget
which the left toobar items are packed into. We can set the sensitivity
on this and get the desired effect.