hid/gtk: Fix styling of menubar area so it is consistent across its width
This relies on some nasty tricks copying GtkStyle's between the menubar
widget and a new GtkEventBox widget we pack the top bar in to give us a
background to paint on. We also adjust the styles of the coordinate read-
out labels and frames to ensure they are legible.

There may still be some inconsistencies, as the coordinate selector button
is themed as if it were a toolbar button, hence the colourings used may
be designed to work with a different background colour than we have.