hid/gtk: Remove redundant (and misplaced) hidgl_flush_triangles() call.
If we were to flush the triangle buffer in show_crosshair(), we should
do it before we set XOR mode on the GL context (assuming the buffer
might be dirty when we are called), OR we should do it after we finish
drawing (assuming we wish to leave the buffer clean when we return).

The location the flush was in before this commit is just plain wrong.

As we currently have flushes of the triangle buffer wrapping the
call to show_crosshair(), just leave the flushing out of this function.

Does this commit message provide an adequate example of simple fix,
complex commit message phenomenon?