undo.c: Add error message where {Save,Restore}UndoSerialNumber() fails
If any operations are added logged for Undo between a call to
IncrementUndoSerialNumber() and call for RestoreUndoSerialNumber(),
those operations will now be in the undo stack with an inconsistent
serial number.

The {Save,Restore}UndoSerialNumber() API is pretty fragile in this
regard, and we should avoid using it where possible. An alternative
might be to implement a "LockUndoSerialNumber()" call which temporarily
disables IncrementUndoSerialNumber(), and a UnlockUndoSerialNumber()
call which restores its function.