undo.c: Add error message where RestoreUndoSerialNumber() will break things
If operations are addded to the undo stack between a call to
IncrementUndoSerialNumber() and a subsequent call for
RestoreUndoSerialNumber(), those added operations will be placed in the
undo stack with an inconsistent serial number.

The {Save,Restore}UndoSerialNumber() API is pretty fragile in this
regard, and we should avoid using it where possible.

An better alternative might be to implement a "LockUndoSerialNumber()"
call which temporarily disables IncrementUndoSerialNumber(), and a
"UnlockUndoSerialNumber()" call which re-enables its increment function.