hid/gtk: Set the LC_NUMERIC locale to "C" rather than "POSIX"
This value is supported on Win32 platforms, whereas "POSIX" doesn't
appear to have any effect.

This was one possible way to fix the "zoom doesn't work" bug on Win32
for locales which use "," as the decimal separator.

DJ fixed the Zoom action to be locale independent in
commit fd5399c67b988f2f7c9d1a0b4ab7c13bc2e95158, which worked around
the problem for this particular action, but I feel it would be best
that we made our locale handling consistent between Linux and Win32
as well.

Affects-bug: lp-843577