Revert "Change change.c to treat text scale as mils"
This reverts commit 9396377717f7c49cf38d753a35e32d004294b09e.

(And fixes what appears to have been the underlying issue).

See commit 2d21a8ad8f38e07b17450450a096f1ce4b3bef0e for a description
of how text scale is defined and used.

If we want to change this in the future, we need to change usage in
many more places than this reverted commit touched.