crosshair.c: Snap to later-tested objects equi-distant with the current snap.
This inequality adjustment means that when an element mark is on-grid, and
the first-snap is a grid location snap, the test against the element mark
will now pass, and mark the snap as a non-grid snap.

This fixes the rather irritating bug where certain footprints, such as SMT
resistors became impossible to grab by their element mark when that mark
landed exactly on a grid point.