action.c: Fixup some syntax so it appears as legal C before macro expansion
"if TEST_FLAG (...) {" does not look like legal C until the TEST_FLAG macro
is expanded. This works because the macro is wrapped with a () like all
good function-like pre-processor macros.

This bad style was upsetting cppcheck, which flagged up a syntax error for
this line. (Obviously it was not expanding macros before checking).