Currently PCB user can draw arcs at limited angles (usually 90 degree step), though there is no such internal limit for arcs angles.
Now the patched tool is expected to work like this:
1.point at the center,
2.the first point of arc
  (up to 270 degrees; sets the arc direction),
3.the second point of arc (the direction is the same as at the end
  of 2.; when the new arc overlaps the arc drawn in step 2., a full
  circle is created);
4.return to 3. until ESC (`F3' moves the starting point
  to end of the latest arc and returns to 3.)

During 3. and 4. Shift-click moves the center of the arc
to the pointer.

Switching from/to Line tool draws chained segments.

HOW TO TEST
Select Arc tool; draw different arcs with lines
pressing different buttons, and `F2', `F3' and `U' keys.


[ Modified by Peter Clifton to suit HID API changes, and remove a few unused variables ]


Signed-off-by: Bert Timmerman <bert.timmerman@xs4all.nl>