From ad2380fdd76dac904c4449f14bb043ed2bc5650c Mon Sep 17 00:00:00 2001 Subject: [PATCH] lines with different clearances are not merged
Problem:
When multiple lines (with same direction, thickness, and clear flags) and moved, they are merged as one line.
Only the clearance of the first line is used for the merged line.
This behaviour is not wished when the lines have different clearances.
Solution:
Verify the lines clearances and only merge lines if they hace the same clearance.

Signed-off-by: Kevin Redon <kevredon@mail.tsaitgaist.info>