Tag physically/logically connected items differently in the "Find" action
This commit restores the semantic meaning of the "found" flag to what it
was before commit 764c3560a722c768a7048f5c70811ec363862882, and uses the
"connected" flag to identify which of the "found" (logically connected)
items are physically connected to the search origin.

The aim here is to differentiate between what an ohm-meter would show
connected with the board as-currently designed, and what it would read
if all rat-lines were implemented.