toporouter.c: Remove un-necessary call to ResetConnections()
This is not required by the call to CollectSubnets() which follows it,
and we don't have any other requirement to clear flags.

Futhermore, as toporouter.c does not setup which flag is being operated
on, the effect of this call will depend on which flag has been left in
"TheFlag" variable in find.c.