Add option to "Select" action, differentiating "FOUND" and "SELECTED" objects
The exising "Select" action included the variant "Select(Connected)", which
selected any object with the "FOUNDFLAG" set.

As we now make a distinction between logical and physical connectivity
(FOUND and SELECTED) respectively, this commit introduces a split into
"Select(Found)" and "Select(Connected)" variants.

There is a change here, in that "Select(Connected)" now acts on objects
with the CONNECTED flag set, (not the FOUND flag as it did previously).