hid/gcode: Add #include for locale.h for setlocale()
For some strange reason, the lack of this include caused a failure to build
on my machine when compiled without optimisation (-O0). With (any)
optimisation turned on, the build worked fine without the include.

Having stripped this down to a small test-case c file caling setlocale, I
discovered that our definition of setlocale and the various LC_* constants
were being pulled in via our include of global.h, which includes libintl.h,
which in turn pulled in locale.h.

I am unsure why this was not the case when -O0 was used to compile.