hid/gtk: Improve mouse event handling
Lets pretend we never leave the drawing window, so we don't get bugged to
click a location when we scroll outside of it during a drag.

Also, use any event coordinates from the scroll event. Never hurts to have
up to date coordinates!

A better fix might be to check whether an implicit grab is in place in the
leave event, if so, we can skip setting have_entered to false in that case.
This may involve tracking whether we have a matching set of button pressed
and button released events.

X11 creates an implicit grab causing events to be delivered to the window
where the a mouse-button was pressed down. Until that button is released,
we have effectively NOT left the drawing widget, even though GDK/GTK
helpfully send us a leave event as we cross its boundary.

If there is an implicit grab when we leave the window, we should mark
has_entered as FALSE when the offending mouse button is released.


Owing to the possible complexitis here (no to mention when you consider
multiple input devices etc..), I'm passing on this implementation for now!